+++ components/overview/BillsOverview.js
@@ -4,7 +4,7 @@ import formatVal from '@/utils/formatVal';
 import billData from '@/utils/billData';
 
 const BillsOverview = ({ data }) => {
-  const { paidTotal, upcomingTotal, dueSoonTotal } = billData(data);
+  const { paid, upcoming, dueSoon } = billData(data);
 
   function billRow(title, value, color) {
     return (
@@ -28,9 +28,9 @@ const BillsOverview = ({ data }) => {
         link="/bills"
       />
       <div className={styles['bill-rows']}>
-        {billRow('Paid Bills', paidTotal, 'green-bg')}
-        {billRow('Total Upcoming', upcomingTotal, 'yellow-bg')}
-        {billRow('Due Soon', dueSoonTotal, 'cyan-bg')}
+        {billRow('Paid Bills', paid.total, 'green-bg')}
+        {billRow('Total Upcoming', upcoming.total, 'yellow-bg')}
+        {billRow('Due Soon', dueSoon.total, 'cyan-bg')}
       </div>
     </div>
   );
