+++ styles/components/filters/Dropdown.module.css
@@ -1,0 +1,69 @@
+.dropdown-div {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+}
+
+.dropdown {
+  position: relative;
+}
+
+.dropdown-button {
+  height: 2.8125rem;
+  border-radius: 0.5rem;
+  border: 1px solid #98908b;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 1rem;
+  padding: 0.75rem 1.25rem;
+  color: #201f24;
+}
+
+.dropdown-button:hover {
+  background-color: #f2f2f2;
+}
+
+.dropdown-button--active {
+  outline: 1px solid #201f24;
+}
+
+.dropdown-button-sort {
+  width: 7.125rem;
+}
+
+.dropdown-button-category {
+  width: 11.0625rem;
+}
+
+.dropdown-menu {
+  z-index: 9999;
+  position: absolute;
+  top: 3.3rem;
+  list-style: none;
+  padding: 0 1.25rem;
+  box-shadow: 0 0 1rem #88888880;
+  background-color: white;
+  border-radius: 0.5rem;
+  transition: all 0.2s ease-in-out;
+}
+
+.dropdown-menu--open {
+  opacity: 1;
+  visibility: visible;
+}
+
+.dropdown-menu--closed {
+  opacity: 0;
+  visibility: hidden;
+}
+
+.dropdown-menu-sort {
+  width: 7.125rem;
+}
+
+.dropdown-menu-category {
+  width: 10.5rem;
+  height: 18.75rem;
+  overflow: auto;
+}
